scenarios/yaml_load/src/backend/requirements.txt
pyyaml
pycgi
chalice